pyproject.toml
[tool.poetry]
name = "malw"
version = "0.1.0"
description = "A Simple Malware Analyzer with Python"
authors = ["0xstvrs"]
readme = "README.md"

[tool.poetry.scripts]
malw = "malw.malw:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.3.2"
python-magic = "^0.4.22"
pefile = "^2019.4.18"
ssdeep = "^3.4"
tabulate = "^0.8.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
pyinstaller = "^4.3"
rope = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
